Use Quasi quoter to parse trusted-certs.crt at compile time.
